Refactor: rename packages to correct casing
