--- src/main/java/squid/DukeIO/TaskStr.java
@@ -1,10 +1,0 @@
-package squid.DukeIO;
-
-
-import squid.Tasks.Tasks;
-
-public class TaskStr {
-    Tasks tasks;
-
-
-}
