+++ src/main/java/squid/constants/EXCEPTIONS.java
@@ -1,4 +1,4 @@
-package squid.CONSTANTS;
+package squid.constants;
 
 public class EXCEPTIONS {
     public static final String INCORRECT_INPUT = "You haven't taught me that trick yet..";
