+++ src/main/java/squid/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 
 public abstract class DukeException extends Exception {
     public DukeException(String message) {
