+++ src/main/java/squid/exceptions/DuplicateTaskNameException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 
 public class DuplicateTaskNameException extends IncorrectInputException {
