+++ src/main/java/squid/exceptions/IncorrectIndexException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 public class IncorrectIndexException extends IncorrectInputException {
     public IncorrectIndexException(String message) {
         super(message);
