+++ src/main/java/squid/exceptions/NotEnoughDatesException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 
 public class NotEnoughDatesException extends NotEnoughInputsException {
     public NotEnoughDatesException(String message) {
