+++ src/main/java/squid/exceptions/NotEnoughInputsException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 
 public class NotEnoughInputsException extends IncorrectInputException {
     public NotEnoughInputsException(String message) {
