+++ src/main/java/squid/exceptions/SquidDateException.java
@@ -1,4 +1,4 @@
-package squid.Exceptions;
+package squid.exceptions;
 
 public class SquidDateException extends DukeException {
     public SquidDateException(String message) {
