+++ src/main/java/squid/io/TaskStr.java
@@ -1,0 +1,10 @@
+package squid.io;
+
+
+import squid.tasks.Tasks;
+
+public class TaskStr {
+    Tasks tasks;
+
+
+}
